requirements.txt
colorama~=0.4.6
termcolor~=2.4.0
PyYAML~=6.0.2
tabulate~=0.9.0
dnspython~=2.7.0